correction d'un bug lors de la lecture de certains fichiers laz qui donnaient des positions de point incorrectes, les positions sont récupérées comme des integers et non comme des long

+++ trunk/VoxeLidarGUIFX/src/main/java/fr/ird/voxelidar/voxelisation/als/LasPoint.java
@@ -68,7 +68,7 @@ public class LasPoint {
      * @param gpsTime
      */
     
-    public LasPoint(long x, long y, long z, byte returnNumber, byte numberOfReturns, int intensity, double gpsTime){
+    public LasPoint(int x, int y, int z, byte returnNumber, byte numberOfReturns, int intensity, double gpsTime){
 
         this.x = x;
         this.y = y;
